Make taxsim.RData
This is the first commit where my plan was to create a single long lookup
file of potential tax rates.  The plan is to develop a function in ret2.R
that takes income adn savings decisinos and returns total tax liabilities.

Currently, taxsim.RData isnt' actually that practical.  The levels that I
chose aren't that useful.  But this shoudl work so I could run it once on
the cluster and that ought to be OK.