Beginnings of cap gais rate
This is the beginning of work that will calculate the marginal rate for
capital gains income separately.  I also added the option to calculate taxes from taxsim to the
taxlookup function, but this doesn't work.

The key thing is that you give taxsim info about what marginal rates it
reports