Simulate life-cycle wages
This takes a draw from the model in Gottschalk and Moffitt 2011, without
time dependent variances in volatility.  It needs to be hooked in to the
rest of the code but seems pretty reasonable

One issue to consider is that in fact you might not want to simulate a
shock growth parameter, since this will really dominate wage growth, and
you wouldn't want your retirement calculator to tell you to save or not
based on wages being affected by this stuff.