Put simulation and switch in ret2.R
There are a few changes here in sim.R as I was playing with it.  The main
issue was that i was thinking abotu whether the alpha in Moffitt and
Gottschalk was a per year or per 10 year thing.

ret2.R has a new way of calculating laborincome, using the simulated
lifetime earnings.  I turned it off by default now because it's weird.
But it's good to use the cell-based thing.

calculate.R now sets up the parameters appropriately for this